backend/requirements.txt
# Core Django packages
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Production server
gunicorn>=21.2.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Static files
whitenoise>=6.6.0
